requirements.txt
Django==1.9.8
django-filter==0.14.0
djangorestframework==3.4.0
psycopg2==2.6.2
